backend/requirements.txt
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
stripe==9.0.0
python-multipart==0.0.6
pymongo==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pyotp==2.9.0
qrcode[pil]==7.4.2
bleach==6.1.0
requests==2.31.0
Pillow==10.1.0